pyproject.toml
[project]
name = "new-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiofiles>=25.1.0",
    "asyncpg>=0.31.0",
    "beautifulsoup4>=4.14.2",
    "deepagents>=0.2.8",
    "fastapi>=0.128.1",
    "langchain>=1.0.8",
    "langchain-anthropic>=1.2.0",
    "langchain-community>=0.4.1",
    "langchain-google-genai>=3.1.0",
    "langchain-ollama>=1.0.1",
    "langchain-openai>=1.1.0",
    "langchain-postgres>=0.0.16",
    "langchain-tavily>=0.2.13",
    "langgraph>=1.0.3",
    "langgraph-cli[inmem]>=0.4.9",
    "openevals>=0.1.2",
    "psycopg[binary,pool]>=3.3.2",
    "pymupdf>=1.26.7",
    "pypdf>=6.4.0",
    "pytest-xdist>=3.8.0",
    "python-dotenv>=1.2.1",
    "python-multipart>=0.0.22",
    "rich>=14.2.0",
    "tavily>=1.1.0",
]

[dependency-groups]
dev = [
    "agentevals>=0.0.9",
    "debugpy>=1.8.17",
    "langgraph-cli[inmem]>=0.4.7",
    "langsmith[pytest]>=0.4.46",
    "pytest>=9.0.1",
    "pytest-asyncio>=1.3.0",
]

[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"
